[KGEN] Re-add support for region parameters in the Elaborator. (#9647)
This patch adds support for region parameters back into the elaborator. It assumes that the inlining of any always_inline calls would have already happened.

There's an interesting constraint that's worth mentioning - we assume that the region parameter is only visible in the scope in which it's defined and scopes below that. Because we use the ParameterUseDefGraph to look up the region referenced, the attribute is only resolve-able in those particular scopes. Subsequent patches will lift this restriction.

MODULAR_ORIG_COMMIT_REV_ID: 726f12e5500717c4558e79709c18d7d5b6820d60