[lit-lang] Implement support for memory-primary by-value arguments. (#9825)
This implements support for declarating and passing memory-primary
values to function arguments, including self.  The most important
piece of this is the introduction of a new 'ByValInMem' argument
convention that models an initialized argument being passed in by
reference, but which expects the caller to take the value in question
and destroy it (when we have destructors) rather than returning it
back out and having the caller destroy it.

On top of that, various call machinery is taught to produce that
convention when a memory-primary value is passed, and taught to
add PointerType as needed.

MODULAR_ORIG_COMMIT_REV_ID: 181300145683f9eebeb5b66789562f8198d71059