[lit-lang] Introduce a new [M]BValue value classification. (#9852)
This introduces a new "borrowed value" classification to represent
an unowned reference to a value sitting in a memory or register
(currently implemented just for memory).  This will be important for
borrowed argument conventions in time, but is immediately useful for
expressions like "a.x.y" where we don't want to materialize an
independent copy of "a" or "a.x" eagerly, but we do materialize it
in the testcase where it was getting dropped.

This makes the RValue categorization cleanly represent a wholely
owned instance of a value, similar to an x-value in C++ terminology.

We'll need a corresponding SBValue model at some point, but right
now we can't implement clone properly for register-primary types.

MODULAR_ORIG_COMMIT_REV_ID: d1a7be929ac35859a6686413a4bc1f2ab581719f