[lit-lang] Allow functions to omit final return if it is unreachable (#9838)
This PR implements a light DCA in `lower-lit-terminators` to determine
whether the endof the function body is reachable. If it is not and a
`return` was not explicitly specified, then we insert a
`kgen.static.undef` to permit this.

Fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: e0394a3e272d6a9918488c09821b252e19a07b0b