[KGEN] Infer the output file extension from the command being invoked (#9888)
Before everything was generated with the .mef extension which made it
a bit hard to work with. Now we infer the extension from the command
being invoked. This means that when we ask lit to generate a doc file
we get a .mt extension rather than .mef.

MODULAR_ORIG_COMMIT_REV_ID: 30b62fd8c5dafe4c514a92c31e158bcc567ed036