[KGEN] Support exporting variadic arguments in function declaration (#9889)
This will allow us to pass around variadic arguments in function calls
from C++ to Lit kernels. This is needed for the upcoming support for
Ops that have variadic arguments like Concat and Slice. Note that there
is only support for exporting variadic arguments and not the input
arguments (it looks like emitSignature is busted for return arguments
anyways, so this can be fixed in a subsequent PR).

MODULAR_ORIG_COMMIT_REV_ID: 2ab80c8928486edc7ba35b0e125254a285df15d4