[MO] Do capture analysis after lowering regions in mo.if (#9892)
When @mdanatg tried to use mo.if for lowering tf.if with tf fallbacks,
he found some issues with captures of the tf context object used by the
fallback.

This PR edits the mo.if lowering to do capture analysis after lowering
the then and else regions. This has the advantage of removing the
complexity from the lowering state I introduced in #9525, and reduces
the amount of context/chain tracking the lowering needs to do.

Contains a small change to a support utility for doing post-lowering
capture analysis, and a small change to TF to properly lower bools to
MO.

MODULAR_ORIG_COMMIT_REV_ID: b99d588b3ade561f8ab74ba883a3c55a63a7332e