[lit-lang] Clean up terminology, use 'memory-only' and 'register-passable'. (#9984)
The model of memory bound types is that they may /only/ be referenced
in memory, they may /never/ be loaded and passed around.  This is
because they need to maintain identity, e.g. for something like an
std::atomic.

On the other hand, we will have an opt-in model that allows types to
be marked "register-passable" which means they don't care about
identity (and are small) thus make sense to pass in SSA registers
instead of going through memory.

This patch renames memory-primary to memory-only, and register-primary
to register-passable consistently.

MODULAR_ORIG_COMMIT_REV_ID: c1a85c34074674828eaeae73addf7877aec74592