[All] Remove `NO_LLCL_RUNTIME` macro (#10004)
The `NO_LLCL_RUNTIME` originally introduced in

[Internal link]
is no longer needed due to

[Internal link]

The current state of affairs is that including `LLCL/Runtime/Runtime.h`
in multiple places (such as in a dynamic library and an executable, or
multiple thereof) is not an issue.

The core sharp edge right now is multiplicity of some singletons like
the TypeID info table (one per executable or dynamic library). In these
cases, the ordering of the link line is the thing that matters most for
the loader determining which symbol references is the One True Winner.
Right now, we're sneaking by and things are "just working" barely in
`tf-api-validator` and friends since they're getting the TypeID info
singleton from `libtf.dylib` and all references route through that
instance rather than the one statically brought in from the "other
static linking" of `LLCLRuntime.a` (such as from
`modular-framework-common`).

MODULAR_ORIG_COMMIT_REV_ID: 760448bf7db6fbfe9c36752eb13a1d62de0fdeda