[KGEN] Persist the requested BindSignature type unless unspecified (#10051)
`simplifyBindSignature` uses `getSpecializedSignature` to compute the
result type of the parameter operator. However, in cases where a more
specialized ParameterEvaluator is used, such as in the elaborator, the
result type will mismatch because the elaborator requests a more
specific type. This change allows `simplifyBindSignature` to use the
requested type if present and the default specialized type with a
vanilla `ParameterEvaluator` if not.

MODULAR_ORIG_COMMIT_REV_ID: bcfa5b339c03dc98ea49d5cf152974a4eec301cd