[lit-lang] Rename LValue to SLValue, and re-introduce LValue. NFC (#10132)
When we support computed LValues, we need to be able to represent
expressions with getters/setters, which are a new form of LValue.
To make room for this, rename LValue to SLValue (SSA LValue, an
LValue that has a known physical address) and introduce a new
LValue class that can be either a physical LValue or a computed
one.

This stubs out but does not implement CLValue.  I considered using
the name "PLValue" for "physical lvalue" but chose not to because
the P in PRValue means parameter, and I thought that naming these
similarly would be very confusing.

MODULAR_ORIG_COMMIT_REV_ID: 4620daccf1bbedfc0536f0a1960fab6d6b68296b