[C API][CAS] Add sourceModelBuffer caching (#9924)
This change adds caching of the source model buffer into a CAS cache.
The CAS cache `modelCASCache` is stored as part of a `RuntimeContext`.
This change also adds a `ModelId`, which is stored in `CompiledModel`.
This is a wrapper for the hash of the cached model buffer and is exposed
to the C API via `M_ModelId`.

This will enable future changes to call `M_compileModel()` with an
`M_ModelId`.
Future changes can also skip compilation if a model already exists in
the `modelCASCache`.

Limitations:
- Currently the `modelCASCache` is hardcoded to be a filesystem
  BlobCache backend.

MODULAR_ORIG_COMMIT_REV_ID: fef1afb140982d73e02d802304d4cef732bf4161