[Stdlib] Simplify the polynomial interface and add more tests (#10173)
This simplifies the polynomial interface by relying on the auto
splatting
behavior of the SIMD type. This means that we can remove a bunch of code
that was in the polynomial functions and just rely on the coefficients
already being SIMD values. We also add tests that check whether the
estring evaluation matches with the horner and polynomial evaluation.

MATH_ORIG_REV_ID: 3063224c5cdb91883992955fed68df4fdce36070