[Cache] Add BlobCache::find that takes a preallocated buffer. (#10134)
This patch adds (and plumbs through) a new `find` implementation that allows the user to pass in a pre-allocated buffer. For the InMemoryBackend, the implementation currently holds onto a ref to the buffer that was passed in for read, to avoid extra copies. This policy decision could be reversed/changed if desired, keeping a copy of the data rather than simply a ref to the buffer. Unfortunately, for the FilesystemBackend, we must copy the file contents into the buffer.

MODULAR_ORIG_COMMIT_REV_ID: 966a6d30df906a0084265348201817b6e7c38267