[lit-lang] Implement DLValues and __setitem__ support. (#10240)
This implements support for DLValue, a dynamic lvalue that consists
of a getter and setter pair.  When accessing the LValue in a read
position we emit a call to __getitem__ and when storing a value
through such an LValue as a ValueDest we emit a call to __setitem__.

There is more that needs to be done here, but stuff is starting to
light up.

MODULAR_ORIG_COMMIT_REV_ID: 69dbbe2eb09ce971296accbc6d554e8f86afc7de