[Mojo] Enable module resolution for both mojo and lit files (#10282)
This enables the resolution of modules with both .mojo and .lit file
extensions. This a temporary solution until we can fully migrate all
files to use the .mojo extension. Both the lit command and CMake will
check if a file exists with the .mojo extension and use that if it does
then it is returned, otherwise the .lit extension is queried.

To demonstrate this, the _CompilerBuiltin has been moved from using the
.lit extension to using the .mojo extension. We will gradually move
files over the next week to use the .mojo extension. Once that's done
we can remove the .lit extension from the resolution logic.

MODULAR_ORIG_COMMIT_REV_ID: df388d75245e73020ed5d661e965de04dfce9d64