[MO] Don't reallocate constant attribute backing data. (#10220)
MOToPrimitives.cpp is responsible for lowering mo.constants to
mgp.buffer.constants, including converting the constant attribute to a
just-an-array-of-aligned-bytes form ready for MEF. This PR makes sure we
don't copy the original attributes backing data when creating a
dense_resource, and makes sure we can reuse dense_resources from an
earlier pass even if it's blob alignment does not match the desired
alignment.

- Push !mgp.aligned_bytes down to !M.aligned_bytes so we can use it to
convey alignment on dense_resources irrespective of underlying blob
alignment.
- Quick syntax tweak to make the alignment part of that type unambiguous
from the size part (I was always forgetting which was which).
- Some helpers for constructing !M.aligned_bytes from dense_resource and
M.array attributes which will look for type annots and do the right
thing for byte size calculation.
 - Try to centralize alignment recovery to those helpers.
- Drop support for dense attributes from MEF since subsumed by our own
attributes and we didn't respect splat anyway.
- Make MEF serialization respect alignment coming from type annotation
if any.

With this we can transition to dense_resource early on, layer on
alignment once we know it, and get the desired alignment on
serialization to MEF, all with zero data copies.

Looks like the upstream M::legalizeDenseElementsForMEF is copy free, but
will check.

MODULAR_ORIG_COMMIT_REV_ID: a5018c42a1bd264cb79fc1335dfe4bcf659c4139