[mojo-lang] Refactor call code to allow capturing IRValues from calls. NFC (#10366)
Emitting a call starts with a set of IRValues which are then
coerced to the expected type and convention of the callee,
after overload resolution.

This coersion can generate code which has observable side effects:
for example, lvalue to rvalue conversion emits calls to getters.
Normally this is fine because we only emit a call once, but in the
case of DLValues, we need to emit a get and set pair, and we only
want the index expressions to be emitted once.

This patch changes the call emission logic to mutate the IRValues
in place to reflect the conversions performed, allowing the setter
to avoid re-emitting the conversions.  This isn't actually used yet
though!

MODULAR_ORIG_COMMIT_REV_ID: 09433ec05ddb389cd7f94c44a02dbd5d63c4c54b