[CACHE] Refactor cachedTransform to enable generic transformations (#10234)
Right now all of the transformations are anchored on an operation,
but the concept of a cached transformation is really useful in many
places (outside of doing something to a specific operation). This PR
restructures the main logic to be agnostic, and adds Operation wrappers
on top.

MODULAR_ORIG_COMMIT_REV_ID: 515fc6da533fc15a130358158a16f606c55930ca