[LLCL] Fix POSIX semaphore check and fix backup semaphore (#10405)
On Linux, we are (always?) falling back to the "Backup Semaphore"
implementation in `LLCL/lib/Support/Semaphore.cpp`, and the wait()
implementation of this semaphore has a bug: counter is not decremented
and so idle threads looping over sema.wait() essentially never block on
the semaphore and end up busy waiting.

The reason why the backup semaphore implementation is used is because
the `check_symbol_exists(sem_timedwait "semaphore.h" ...)` fails even
when the symbol exists. I observed this on my Ubuntu 22.04 machine as
well as Ubuntu 20.04 docker image that I used to build Triton. Adding
pthread as a required library fixes this.

MODULAR_ORIG_COMMIT_REV_ID: 7dd9b2f44151303fd747ea9e598e45186174a515