[KGEN] Add bit on `RegionAttr` to store if the region is isolated from above. (#10401)
This patch adds a bit to cache the SSA isolation of a region parameter. This will be useful when we handle region parameters in the elaborator because this is the main thing that would prevent verification of a FuncOp expansion tree node - so we don't have to loosen the verifier as much.

MODULAR_ORIG_COMMIT_REV_ID: 273e417c48c45aaf6afab95ced0c41eaaa9e5fa4