[Kernels] Disable generic transpose for sizes covered by specialized impls. (#10449)
Currently `transpose` has 4 implementations, 3 of which specialize on 4,
8, and 16 sizes, and the fourth one is generic. This means that for
sizes 4, 8, and 16 transpose has 2 implementations.

This results in a elaboration space explosion when we generate matmul
kernels: these kernels extensively use `unswitch` which means that we're
quickly multiply the number of candidates. The final number of
candidates for matmul tiled is 256, which goes down to 1 with this
change.

This significantly improves compile time for `MatmulKernels.lit` and
`NN.lit`:
```
  MatmulKernels.lit:   95s   ->   7.7s
  NN.lit:            16.7s   ->  13.7s
```

LINALG_ORIG_REV_ID: f96e65fc8550c784a52da29b83afddd77a93a11f