[lit-stdlib] Fix brittle threading heuristic in gather-sum kernel (#10465)
Makes the min task size a function of the number of elements copied,
rather than an arbitrary number of rows. This is an NFC for
dlrm-rm1-multihot since the tasks size is chosen based on the dims of
that model.

NN_ORIG_REV_ID: 72726637cb0bb19f38bfe37c7858278791c439dd