[KGEN] Parallelize LLVM codegen and object creation (#10235)
Our current compilation pipeline is similar to LTO, where we
essentially compile everything in the program together and
emit one giant object file. This doesn't really scale to the
programming model of lit, where a *lot* of things are parametric
and lots of IR is generated.

This PR slices up the LLVM module and performs code generation
in various different threads, linking all of the sliced object files
into a single archive. For now, the splitting is done at the
llvm::Module level for simplicity (because LLVM already has some
facilities for cloning and splitting modules), but realistically
we should move the splitting higher up the stack (which would bring
various benefits along with it).

The splicing is done per exported function, with all of the internal
dependencies cloned as necessary. This leads to duplication in
the generated archive, but this should mostly be a temporary step
while we continue to depend on LLVM for interprocedural optimization.
Even though we can't yet get finer grained splitting, even splitting
at the export boundary can bring dramatic speedups in codegeneration.
For example local compiles of NN.lit and ElementWise.lit get cut
in half.

MODULAR_ORIG_COMMIT_REV_ID: ae39c40c76e4e3ffd6e833bb99a79dd7d3993b9c