[kgen] Split header generation from archive generation (#10462)
This PR moves header generation to its own command, emit-header,
instead of grouping it with emit. This also updates add_kgen_kernel to
generate two targets, `target` and `target-headers`. The latter is
intended to be used in the DEPENDS section, while the former is
used solely in the link libs.

The main benefit of this is that we no longer serialize the build
process by waiting for kernels to finish compiling, we can build
much more of the code base in parallel now.

MODULAR_ORIG_COMMIT_REV_ID: 1c31bed5d3dceb109e7cabfefbd4c2eb970e9df7