[KGEN] Fix printing of `param.fork`. (#10491)
`param.fork` assumed that it could always resolve its element type when
printing - this is not always the case. If you can't resolve the element
type, just print the type of the value itself. This is obviously not
perfect, but it allows the elaborator to avoid a crash while debugging.

MODULAR_ORIG_COMMIT_REV_ID: 71fe072e37e91c16f623717d9e0964291f273b39