[Kernels] Optimize gather-reduce kernel performance for ARM (#10488)
Fixes [Internal link]

Previously the gather-reduce kernel was ~2x slower than the vanilla
gather kernel on ARM (e.g. M1). This change accumulates in a simd
register instead of the output buffer which results in improved
performance on both M1 and Skylake architectures. With this
optimization, the gather-reduce kernel is strictly faster than the
gather kernel on M1 (~1.3x) and 2-3x faster than the previous
implementation of the gather-reduce kernel. It also results in up to a
1.2x speedup on Skylake. Detailed figures in comments.

NN_ORIG_REV_ID: b971b985c34fa317f4ac10f04518c3e893893542