Revert "[mojo-lang] Refactor call code to allow capturing IRValues from calls. NFC" (#10476)
This reverts [Internal link] I think there is a different
approach with less invasiveness that won't need this, and I'd like to
keep the internals simple and more 'functional'. I also don't have time
to really drive this to the ground (and it is low priority), so I'll
file an issue and come back to this later.

MODULAR_ORIG_COMMIT_REV_ID: d598d3dfaa853425caf70422f9c4bb9ec6d2fa40