[CAS] Make generic DylibBlobCacheBackend and use for S3 backend (#10176)
This introduces a generic DylibBackendStub as part of the Cache module,
that can be used to load any BlobCacheBackend (with dlopen)
implemented in a shared library.

This stub is now used to access the S3 BlobCacheBackend implementation,
which is still in `CloudInfra/CAS` but all of the dylib stuff has been removed
from that module.

MODULAR_ORIG_COMMIT_REV_ID: 06b31ee3b2aee9e11627e966aceb9c29b756e392