[mojo-lang] Start adding support for `__init__` (#10540)
One weirdness of Mojo vs Python is that it uses `__new__` for structs
instead of `__init__`.  It is more uniform to support the thing that
Python user's expect, and now that we're getting plumbing for it,
we can switch to __init__ over time.  I'd like to eventually
completely eliminate __new__ on structs, as it is a redundant
concept that `__init__` can directly cover.

This patch adds support for declaring and type checking `__init__` on
memory-primary types.  It does not support invoking them or on
register primary types yet.

MODULAR_ORIG_COMMIT_REV_ID: 2a1017eb35c62673babab74876347f07c1dad031