[mojo-lang] Add initial support for validating doc strings (#10513)
As we start more pervasively writing doc strings for the mojo
APIs, it's quite useful to know when you are steering outside of
the expected doc string format (other than generating docs, and
checking to see if things are behaving correctly). This PR intends
to assist in this by diagnosing many common issues as warnings,
which nicely show up in the IDE when editing the strings.

For now the validation is an opt-in feature, because we need to
understand the cost of always doing it (if we wanted to) and also
for the fact that a lot of warnings would be emitted for our current
doc strings.

MODULAR_ORIG_COMMIT_REV_ID: 6b0569675f27a38a4d9f2050a4ce52a5352bdcf9