[mojo-lang] Rename PRValue -> PValue, add emitBValue, NFC. (#10688)
This generalizes PRValue to be able to represent BValues as well,
and clarifies that it can hold memory-only types as well as
register_passable ones.  To reflect this, rename it from PRValue
to PValue.

This is important to make borrowed values happen and should also
be a step towards enabling memory-only parameters.

MODULAR_ORIG_COMMIT_REV_ID: 8215a5994d61eebafadb16db92d1722fd94e02b1