[mojo-lang] Implement new state for ValueDest, NFC. (#10698)
Direct initialization of memory-only buffer, and calls into a
memory-only buffer will directly initialize that provided buffer.
A future patch needs to know the difference between that buffer
taking ownership of that value vs being emitted into a temporary
for a shadow copy - which happens if there is no provided buffer
or if the buffers disagree.

This introduces the machinery for this in a way that doesn't change
behavior but puts the asserts in.

MODULAR_ORIG_COMMIT_REV_ID: cac062a23368f2f3078b617edbeec9abeed4e3ae