[mojo-lang] Rework `emitResult` to be internally type consistent. (#10700)
`emitResult` into a ValueDest has always been a bit tricky: if there
is an lvalue in the ValueDest, the expression being emitted transfers
ownership into that lvalue.  However, if not, it returns ownership
back to the expression caller.

We have a clean way to model this with BValue vs RValue's, but the
existing hooks assume that an 'emitRValue' call would return an
RValue, so we couldn't reflect that properly.

This patch reworks the world so that emitting into a ValueDest that
takes ownership of the value returns a borrow of that value, instead
of redundantly returning ownership.  This eliminates a footgun and
also catches some value copies that were missing previously.

While reworking this, thread some ExprContext's into the new RValue
emission logic.

MODULAR_ORIG_COMMIT_REV_ID: 31b8ca8a19b1743103fa47aa84cb7143847e4ff5