[mojo-parser] Throw away cached modules whose dependencies are not also cached. (#10665)
This should fix the issue with parser caching that we've been seeing in CI. Basically, what was happening was that the dependencies of a cached module were not also in the cache, but the cached module is being marked as fully resolved, and so its dependencies are not resolved. This results in really odd errors when there's an import cycle.

This patch 'fixes' the issue by simply throwing away the cached module if the module's dependencies are not also in the cache. This is clearly undesirable, but the full fix will take more time - #10657 is filed for a follow-up.

MODULAR_ORIG_COMMIT_REV_ID: 6913a0d4fb276ab45d8c05e95386934d1afc1c0d