[Cache] Move generic CachedTransforms out of the CacheDialect lib, NFC. (#10677)
This patch moves the parts of CachedTransform that don't rely on MLIR out of the CacheDialect library so that they can be used in places we don't want to have a dependency on MLIR. It also updates the TransformCacheKey to use a BufferRef rather than a StringRef to avoid confusion, as StringRef keys usually are *not* hashed, whereas the TransformCacheKey *is* hashed. This required small updates in other parts of the code to conform to the new type.

MODULAR_ORIG_COMMIT_REV_ID: cc86b9e3ae390a8146bb7eeba654e66449872a7b