[GenericML] Split `ForeignBufferInterface` globals into its own lib (#10714)
As part of the multi-tenancy work, one thing we're struggling with right now is
the runtime globals crossing `.so` boundaries between `modular-framework-common`
(where the runtime is created) and then getting used in framework-specific
`.so`.  This leads to the infamous "mismatched runtime" canary assert we have,
but it really means that these two `.so` are accessing through different
globals since they each had their own copy.

To make this behavior very intentional and obvious, we need to isolate all of
the globals that come into play during model compilation and execution in our
runtime.  In order to avoid a monolithic patch, this patch is aimed at just
isolating the globals with foreign buffer interface from `GenericML/Support`.

Split the global into a separate shared library: `GenericMLSupportGlobals` and
then call through that from `ForeignBufferInterface.h`.  We'll do a similar
thing with `LLCLRuntimeGlobals`, but that one is a bit more involved and can be
decoupled from this.

Add linking with `GenericMLSupportGlobals` as required everywhere.

Co-authored-by: Mark Shields <mbs@modular.com>
MODULAR_ORIG_COMMIT_REV_ID: 8bf2f354f9fb524fe083fd46cf17b7ffecef425c