[LLCL] Split globals into separate library (#10715)
As part of the multi-tenancy work, one thing we're struggling with right now is
the runtime globals crossing `.so` boundaries between `modular-framework-common`
(where the runtime is created) and then getting used in framework-specific
`.so`.  This leads to the infamous "mismatched runtime" canary assert we have,
but it really means that these two `.so` are accessing through different
globals since they each had their own copy.

To make this behavior very intentional and obvious, split the global definitions
for symbols like the `nextRuntimeIndex` and `allRuntimes` into a separate
library: `LLCLRuntimeGlobals`.  It is expected and desired behavior for an end
executable to link exactly *once* with `LLCLRuntimeGlobals` so that we avoid the
One Definition Rule.  The behavior before, unfortunately, was that we had many
copies of `LLCLRuntime.a` flying around, each with their own globals (one per
.so or executable), and this leads to chaotic behavior in accessing the right
global since we don't have a "forcing point of access" through a particular
symbol from a particular point (library) for example.

Note:
- We could make the amount of exported symbols slightly smaller for
  `LLCL/Runtime/Globals/CompactRuntimePtr.cpp` to only export the singleton
  maker and some way of getting the next runtime index.  But this is working
  fine for now and isn't causing any issues.  For symmetry though with our other
  split of globals, this would be a nice to do.
- There are other globals still present in the codebase: in `TimeProfiler.cpp`
  that are not accounted for in any special "globals" library.  However, these
  thread local globals for the thread profiler contexts aren't in the way of
  multi-tenancy support, so we don't touch them for now.  When we get to
  supporting multi-tenancy of different ISAs, I think we'll want to do this
  work.

Co-authored-by: Mark Shields <mbs@modular.com>
MODULAR_ORIG_COMMIT_REV_ID: 22f052449c5d3af501098b5e18fd9654b768096f