[KGEN] Add `concretizeModule` function to KGENCompiler. (#10685)
This function takes the module and constructs a default pass manager using its context, then runs all the passes from pre-elaboration, elaboration, and post-elaboration. It does not split the pipeline at all - that could be done if profiling indicates we need it, but for now, start with a coarse-grained function.

This function is currently unused, but it will be used in subsequent patches by the ORC.

MODULAR_ORIG_COMMIT_REV_ID: 50bec173ce389448f44714a8efccd35ec28e40b1