[KGEN] Refactor `ObjectCompiler`, NFC. (#10686)
This patch attempts to minimize the amount of state the ObjectCompiler holds. It moves the ModuleOp out of the object compiler so a single instance can be used for more than one module. The previous method was a bit silly, having a single instance of the ObjectCompiler per module is not really a meaningful/useful requirement, and this allows the user to pass in different symbol tables and exported symbol sets to the same object compiler.

MODULAR_ORIG_COMMIT_REV_ID: a619f10f636dec58b91686f8b477d3befb03cc25