[mojo-lang] Change `__clone__` to take self by-borrow. (#10785)
Early in Mojo's life we only had by-copy and by-reference, and you
can't pass something by copy to a clone method, because clone is
what does the copy!

Today we have the borrowed convention, which allows passing a
reference to a value without taking ownership of it, which is the
correct convention to use for `__clone__` (and effectively everything
else).  Switching `__clone__` to use this simplifies the compiler and
enables clone to work for register-passable values as well, closing
a massive hole in the copy constructor system we had.

Doing this also exposes more opportunities to Mem2Reg to promote
values, since it now can see the loads more plainly.  This in turn
exposed some bogus/harmless use-of-uninit variable warnings in
Image.mojo, Math.lit, and Vector.lit which I worked around by
tweaking the code.

BUFFER_ORIG_REV_ID: 18813da62c65de75a11b14135b63a619b4df240c