[KGEN] Pull 'layer' abstraction out of the ExecutionEngine, NFC. (#10840)
This patch pulls the 'layer' abstraction out of the ExecutionEngine and
generalizes it to be able to contain many layers. This will allow
clients to construct Layer objects outside the ExecutionEngine, and add
them as needed. This step is important to avoid layering problems when
we start adding layers for the ObjectCompiler, the elaborator, and even
the parser, since the KGEN compiler internally uses the ExecutionEngine.

MODULAR_ORIG_COMMIT_REV_ID: 6e4f00ca50cb1dc58c6f67ba3a14d3b5f35af747