[KGEN] Use a more fine-grained version for the KGEN caches (#10847)
The KGEN caches currently use the MODULAR_VERSION when building
the cache directories. This is very conservative, but also extremely costly
now that we heavily utilize the cache during the compilation process.

This PR adds a new KGEN_VERSION to be used when building caches,
which allows for preserving the cache more effectively. For example,
only 2 commits out of the last 25 (at the time this commit was written)
would have triggered a change to the KGEN version.

MODULAR_ORIG_COMMIT_REV_ID: 01e8924b566942216307e015fde06aeabc18a4f5