[Stdlib] Improve elementwise kernel thread splits (#10951)
When parallelising the elementwise kernel we check the size vs a static
grainsize just like eigen and pytorch. With 1D kernels this is simple as
we just check the numelements. However for the ND kernel we are only
parallelising over the outer dimensions, so we almost never thread.

Instead this computes the number of threads based on all elements while
still parallelising over the first rank -1 dimensions. I checked and
this is the same behaviour as pytorch.

ALGORITHM_ORIG_REV_ID: ca6b41714a4bfba57a77c495ffc95b1a898be181