[mojo-lang] Improve QoI when using dynamic values in param context. (#10981)
It isn't valid to use a runtime value when in a parameter expression,
for example in a type specification.  This patch consolidates the
various complaining about that problem into a specific hook and gives
an explanation why the entire expression is in a parameter context
instead of confusingly complaining about the nearest parent context.

This resolves Issue #7507.

MODULAR_ORIG_COMMIT_REV_ID: aa8fd55ab118f71de7fec5609394fd4720adbe63