[mojo-lang] Emit pack inference failure warning (#11047)
Emit a diagnostic when a non-concrete value prevents parameter inference
for functions that take pack arguments.

This is a simple implementation that allows for a warning to be emitted.
I also considered a different approach, where I added a member to the
`OverloadFitness` struct for storing pairs of `(location, message)`
that represented any additional diagnostics from inference failures.
The benefit of that approach is that the diagnostic could be made to
only be output when overload resolution actually failed to find any
valid candidates, and it could be output as a note attached to the error
diagnostic "invalid call to 'f'". This would mimic how the "function
declared here" note is output.

In practice, however, I couldn't find any examples in our test suite
that necessitated per-evaluation parameter inference failure messages,
so it didn't seem wise to increase the size of the `OverloadFitness`
struct for such a feature.

MODULAR_ORIG_COMMIT_REV_ID: 15dd8e6d8cee9cda3d2b25ec37b24f997208bf33