[KGEN] Add `KGENCompilerLayer` and adopt it in `mojo`. (#10923)
This patch adds a KGENCompilerLayer that runs all the passes in the `populateElaborateModulePasses` pipeline on the provided module, then forwards it for codegen to subsequent layers of the ExecutionEngine.

This patch exposed an issue (#10920) - basically we don't have a mangler/demangler for mojo function names and it'd be *really* nice to have one.

This patch does not adopt the new layer in `kgen` because `kgen` has a bunch of other functionality, including emitting ASM/LLVMIR that is outside the scope of this patch.

MODULAR_ORIG_COMMIT_REV_ID: 54982e5c0bd56dd8c04baabb25ac9cf78edac66c