[Support] Use .native() for portable TempFile path (#11061)
There are a couple of places where we create a path for a `TempFile`
using `std::filesystem::path`'s `.string()` member function, which is
not portable to Windows.
This change rectifies that by using the portable `.native()` member
function instead.

There is one more usage of `path.string()` as an argument to
`TempFile::create()` which is already fixed in #10984.

This change came as a result of [the
[Internal link]
in #10984.

MODULAR_ORIG_COMMIT_REV_ID: b43fdcd89e81eaf89875d161997bc6816cbf4a05