[LLCL] Switch from using a yeild statement to an isb statement (#11123)
This switches the waiter to use the isb assembly instruction on arm64
(see
https://developer.arm.com/documentation/ddi0596/2021-06/Base-Instructions/ISB--Instruction-Synchronization-Barrier-).
This is the closest equivalent instruction to pause on x86. The prior
code performed a yeild which was effectively a noop. According to
https://mail.openjdk.org/pipermail/hotspot-dev/2021-August/054033.html
"At Amazon experiments with the ISB on Graviton 2 showed it can create a
small delay without consuming ALU resources. It is more reliable than
NOPs. ".

it is also what databases, the C++ library, and rust does.

References:

-
https://github.com/aws/aws-graviton-getting-started/blob/main/perfrunbook/optimization_recommendation.md#optimizing-synchronization-heavy-optimizations
-
https://github.com/rust-lang/rust/commit/c064b6560b7ce0adeb9bbf5d7dcf12b1acb0c807
-
https://stackoverflow.com/questions/70810121/why-does-hintspin-loop-use-isb-on-aarch64
- https://bugs.openjdk.org/browse/JDK-8258604
- https://bugs.mysql.com/bug.php?id=100664

MODULAR_ORIG_COMMIT_REV_ID: 682ed719338ff4ac56625d8f08e23c5ce494e330