[mojo-lang] Move variant-insertion for error handling to parser. (#11159)
This shifts the responsibility for introducing variants for throwing
function returns into the parser.  This simplifies and clarifies
the responsibility of the SemanticCF lowering pass so it doesn't do
weird lookup of the Error type, removes two whole-module passes
from it, and otherwise simplifies it.

This also fixes issues where we couldn't take the address of a
throwing function, and where propagation through an else of a
try block was handled wrong.  This fixes #11154

MODULAR_ORIG_COMMIT_REV_ID: 25fcbde650e6c74b12c294dae3ec2dd96a0d56ba