[mojo-lang] Test missing arg/param name, NFC (#11257)
The Mojo parser emits an error diagnostic when an argument does not
have an identifier, but that error diagnostic is not checked in any of
the existing tests. Add a test for this diagnostic.

The diagnostic itself could be improved: it uses the term "parameter"
regardless of whether a parameter or an argument is being parsed. The
tests added in this commit could be used to test any future improvement
made that would use the correct terminology in the diagnostic.

MODULAR_ORIG_COMMIT_REV_ID: f5b14821c646fa21c218563f94cd3231584a3f50