[MOJO] Asynchronous elementwise kernels (#11208)
We add AsyncTaskGroupPtr for tracking the storage and completion of a
known number of Coroutines representing tasks. When all tasks have
completed a given OutputChainPtr is marked as ready, and the
AsyncTaskGroupPtr deletes its underlying AsyncTaskGroup.

An OutputChainPtr may also be 'traced' with a given label. That time
profile entry (under kOther level 1) will begin on the call to trace,
and end when the output chain is completed. This mechanism temporarily
replaces the use of the Mojo Trace until we can rework it to be async
compatible.

Move all the elementwise kernels to this scheme, both for async
execution and for tracing.

MODULAR_ORIG_COMMIT_REV_ID: 6401ac5c1d022eeb858befafa2b63d139bc20d4d