[KGEN] Compute the ParameterUseDefGraph once per GeneratorOp. (#11308)
We want to use the ParameterUseDefGraph as a symbolic graph, where we compute it once per Generator (really, once per DeclInterface) and re-use that for every object that is specialized *from* it. This patch does the simple version of this by simply storing ParameterUseDefGraph objects in a map next to the `topLevelTrees` map. This is not great for a few reasons: (1) This will be tricky to parallelize without locking, and (2) as it currently exists, this is extra state generated *during* elaboration that's not stored on the tree - we should fix this by introducing a ParametricTreeNode (or similar) that can represent nodes with ParameterUseDefGraphs.

MODULAR_ORIG_COMMIT_REV_ID: d4a99a4fea1d9a2141ed846bb7290f78e2e7af6a