[mojo-lang] Parse `*Ts` pack as a type expression (#11431)
Prior to this change, pack arguments were parsed in a "hacky" way in
Mojo: if an asterisk appeared before the type annotation expression, the
parser would bend over backwards to interpret that expression as a pack
type, instead of its actual parsed type. That is, the type of `*args` in
`*args: *Int` would be `Int`, and special type conversion would occur
based on a hacky `bool isPack` ("did we see an asterisk?") condition.

This change introduces `star_expression`, which is named after the
corresponding grammatical element from
https://docs.python.org/3/reference/grammar.html. In the same way that
https://peps.python.org/pep-0646/#change-2-args-as-a-typevartuple treats
a `star_expression` in a parameter list as meaning "unpack the type
sequence that follows," this commit changes Mojo to parse
`star_expression` in argument lists, and to treat the type expression as
a pack. This obviates the need for any `bool isPack` hackiness, since
the expression is appropriately typed.

MODULAR_ORIG_COMMIT_REV_ID: b5417cd5f6552feac5279feff5f8d049a1d9d09b