[KGEN] Demangle mojo signatures. (#11408)
This patch takes a mangled function name and parses the signatures into
proper MLIR types. Note that this may fail in strange ways if passed in
a post-elaboration function name because of the way we handle parameter
values being mangled in. This patch doesn't handle parameter values
because that is in the process of changing, and so any parsing would
become outdated quickly.

MODULAR_ORIG_COMMIT_REV_ID: 9d2b742cab9be807c493c43e67ac0cb0143323e1