[All] Spelling corrections, NFC (#11435)
I ran a spelling checker over the codebase and fixed most of the
suggestions. This is obviously pretty low-priority, but it was bothering
me and it was pretty quick to do, and I figure it's an improvement.

MODULAR_ORIG_COMMIT_REV_ID: e2ec5680bce94a33b70da7479ca42f20576ebb63