[mojo-lang] Enhance VarLetDeclOp/LetRegDeclOp to track destructors. (#11651)
This enhances the mojo frontend to look up the `__del___` member for
a type when creating a var op node, and attach it to the node when
present.  This will allow a later dataflow pass to insert calls to
the destructor as needed.

Note that this is using `__del___` with an extra _ because I don't
want to distrurb any existing uses of `__del__` in the codebase,
I will switch this over to the proper name when things are baked.

MODULAR_ORIG_COMMIT_REV_ID: 38a7b1c30d340100dd9ceea93843ec207fd8a246