[KGEN] Inline leaf-to-root in parametric inliner (#11616)
This partially addresses
[Internal link]

This PR introduces an InliningGraph datastructure to Inlining.cpp that
computes a leaf-to-root order for inlining. This will temporarily
regress compile times because it is not parallel. That will come in the
next PR.

This PR also fixes yet another bug with the mangling scheme exposed by
the reverse order of the inlining (the inline-ees tend to be much
bigger), by reworking the name uniquing scheme entirely to use a max
subscript.

MODULAR_ORIG_COMMIT_REV_ID: 909fdea9a9f6da32d5aea393fa4390e4d09487a2